move ConfigMananger to Utility
