+++ src/Utility/CMakeLists.txt
@@ -1,4 +1,5 @@
 aux_source_directory(${CMAKE_CURRENT_SOURCE_DIR}/src UtilitySrc)
-add_library(vgg_utility INTERFACE)
-# target_sources(vgg_utility PRIVATE ${UtilitySrc})
-target_include_directories(vgg_utility INTERFACE ${CMAKE_CURRENT_SOURCE_DIR}/interface)
+add_library(vgg_utility)
+target_sources(vgg_utility PRIVATE ${UtilitySrc})
+target_include_directories(vgg_utility PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/interface ${VGG_CONTRIB_JSON_INCLUDE})
+target_include_directories(vgg_utility PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/include)
